final fix, now MVC works in the exercise
